core/requirements.txt
asgiref==3.8.1
attrs==25.2.0
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.19
django-filter==24.3
django-mail-templated==2.6.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
exceptiongroup==1.2.2
flake8==7.1.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
